Coding

+++ src/iVM.Model/Models/EventOccured.cs
@@ -5,7 +5,7 @@ namespace iVM.Model
   public class EventOccured: Event
   {
     public DateTime Date { get; set; }
-    public long Mileage { get; set; }
+    public decimal Mileage { get; set; }
     public uint Expense { get; set; }
     public uint Profit { get; set; }
   }
